Add files via upload
This program is a password hashing problem. You must enter a string, the program will transform it to upper case, and then will print the encoded value. It will ask if you wish to enter a value to decode, if you enter the encoded value given, you will see the original string entered in upper case.